chore: move the apply-resource file to the goals folder for constency with where the other goals are located
