+++ lib/goals/apply-resources.js
@@ -18,12 +18,12 @@
 
 'use strict';
 
-const services = require('./services');
-const routes = require('./routes');
-const deploymentConfig = require('./deployment-config').deploy;
-const secrets = require('./secrets');
-const ingress = require('./ingress');
-const configMap = require('./config-map');
+const services = require('../services');
+const routes = require('../routes');
+const deploymentConfig = require('../deployment-config').deploy;
+const secrets = require('../secrets');
+const ingress = require('../ingress');
+const configMap = require('../config-map');
 
 module.exports = (config, resourceList) => {
   const mappedResources = resourceList.map((r) => {
